似たようなアニメーションをするクラスを共通化
